backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
langgraph==0.2.39
langchain==0.3.7
langchain-openai==0.2.8
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
mysqlclient==2.2.0
gunicorn==21.2.0